created LoginPage component and moved that logic from app/page.jsx
